Initial Proof of Concept for debugging of Child Processes
Current code has debug always turned on for worker (child) processes.
This is controlled by the debug flag in lib/ewd.js

Enhanced ewdMonitor's app.js and index.html to display a debug button
linked to a local instance of node-inspector (running on port 8181). Any
new child processes created via the interface don't show up correctly
until a full refresh is performed due to the passing of only the pid to
the function and no obvious access to the underlying data structure.